backend-fastapi/requirements.txt
# FastApi
uvicorn
fastapi
python-multipart

# Async
httpx

# Encryption
python-jose[cryptography]
bcrypt

# MongoDB & Redis
beanie
motor
redis

# DB Mocks
testcontainers

# Tasks
dramatiq[redis]
apscheduler

# GCP
google-cloud-storage

# Schemas
pydantic[email]>=2.4
pydantic-settings
